+++ packages/ui/FloatLayer/index.js
@@ -1,5 +1,0 @@
-import FloatLayer from './FloatLayer';
-import FloatLayerDecorator from './FloatLayerDecorator';
-
-export default FloatLayer;
-export {FloatLayer, FloatLayerDecorator};