+++ packages/ui/FloatingLayer/Scrim.js
@@ -1,8 +1,8 @@
 /* //TODO: JSDOC revisit
- * Exports the {@link ui/FloatLayer.ScrimBase} component and {@link ui/FloatLayer.ScrimLayer}
- * components. The default export is {@link ui/FloatLayer.ScrimBase}.
+ * Exports the {@link ui/FloatingLayer.ScrimBase} component and {@link ui/FloatingLayer.ScrimLayer}
+ * components. The default export is {@link ui/FloatingLayer.ScrimBase}.
  *
- * @module ui/FloatLayer/Scrim
+ * @module ui/FloatingLayer/Scrim
  */
 
 import kind from '@enact/core/kind';
@@ -11,18 +11,18 @@ import React from 'react';
 import css from './Scrim.less';
 
 /**
- * {@link ui/FloatLayer.ScrimBase} provides an overlay that will prevent taps from propagating
+ * {@link ui/FloatingLayer.ScrimBase} provides an overlay that will prevent taps from propagating
  * to the controls that it covers.
  *
  * @class ScrimBase
- * @memberOf ui/FloatLayer
+ * @memberOf ui/FloatingLayer
  * @ui
  * @private
  */
 const ScrimBase = kind({
 	name: 'ScrimBase',
 
-	propTypes: /** @lends ui/FloatLayer.ScrimBase.prototype */ {
+	propTypes: /** @lends ui/FloatingLayer.ScrimBase.prototype */ {
 		/**
 		 * Types of scrim. It can be either `transparent` or `translucent`.
 		 *
@@ -56,18 +56,18 @@ const ScrimBase = kind({
 });
 
 /**
- * {@link ui/FloatLayer.ScrimLayer} is a layer which adds scrim behind the children.
+ * {@link ui/FloatingLayer.ScrimLayer} is a layer which adds scrim behind the children.
  * Children will always have one higher z-index than scrim.
  *
  * @class ScrimLayer
- * @memberOf ui/FloatLayer
+ * @memberOf ui/FloatingLayer
  * @ui
  * @private
  */
 const ScrimLayer = kind({
 	name: 'ScrimLayer',
 
-	propTypes: /** @lends ui/FloatLayer.ScrimLayer.prototype */ {
+	propTypes: /** @lends ui/FloatingLayer.ScrimLayer.prototype */ {
 		/**
 		 * z-index of the layer. Scrim will have the z-index of layer, and the children will have
 		 * one higher index to display on top of Scrim.