+++ packages/ui/CHANGELOG.md
@@ -6,7 +6,7 @@ The following is a curated list of changes in the Enact ui module, newest change
 
 ### Added
 
-- `ui/FloatLayer`, `ui/FloatLayerDecorator` components
+- `ui/FloatingLayer` module with `FloatingLayer` and `FloatingLayerDecorator` components
 - `fit`, `noAnimation` props added to `ui/TransitionBase`
 - `onHide` props added to `ui/Transition`
 - LESS mixins from `@enact/moonstone` that are general purpose and can be utilized by various UI